# ACPApi/Models/ClienteModel.cs
namespace ACPApi.Models
{
    public class ClienteModel
    {
        public int Id { get; set; }
        public string Nombre { get; set; }
        public string Apellido { get; set; }
        public DateTime FechaNac { get; set; }
        public int Edad { get; set; }
    }
}